start breaking up the unit test megafile
