+++ hub/test/src/index.js
@@ -1,0 +1,10 @@
+import { testAuth } from './testAuth.js'
+import { testServer } from './testServer.js'
+import { testDrivers } from './testDrivers.js'
+
+
+
+testServer()
+testAuth()
+testDrivers()
+
